fixed casing bug, add custom 404 pages
