+++ app/[domain]/[slug]/page.tsx
@@ -1,12 +1,9 @@
 import { notFound } from "next/navigation";
 import { getPostData } from "@/lib/fetchers";
-import { unstable_cache } from "next/cache";
-
 import BlogCard from "@/components/blog-card";
 import BlurImage from "@/components/blur-image";
-
-import { placeholderBlurhash, toDateString } from "@/lib/utils";
 import MDX from "@/components/mdx";
+import { placeholderBlurhash, toDateString } from "@/lib/utils";
 
 export async function generateMetadata({
   params,
